requirements.txt
appdirs==1.4.0
dj-database-url==0.4.2
Django==1.10.5
django-crispy-forms==1.6.1
djangorestframework==3.5.4
gunicorn==19.6.0
olefile==0.44
packaging==16.8
Pillow==4.0.0
psycopg2==2.6.2
pyparsing==2.1.10
six==1.10.0
whitenoise==3.3.0
django_cors_headers==2.0.2
djangorestframework-jwt==1.9.0
PyJWT==1.4.2